<refactor>[front] refactor component releted to home
- rename component from <Loadmore~> to <~SliderWindow> or <PostWindow>
